requirements.txt
<<<<<<< HEAD
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
numpy==2.1.3
outcome==1.3.0.post0
pillow==11.0.0
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
=======
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
numpy==2.1.3
outcome==1.3.0.post0
pillow==11.0.0
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
>>>>>>> 0b1131294f2072f854476bf44c3894b8a2dbbf1b
